+++ src/main.tsx
@@ -1,16 +1,23 @@
-import React from 'react';
-import { useState } from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import Reporte from './Pages/Reporte';
-import './index.css';
-import { createBrowserRouter, RouterProvider, Navigate } from 'react-router-dom';
-import ListadoReportes from './Pages/ListadoReportes';
-import InfoReporte from './Pages/InfoReportes';
-import Login from './Pages/Login';
-import { UserProvider, useUser } from './Pages/UserContext';  // Importa UserProvider y useUser
+import React from "react";
+import { useState } from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import Reporte from "./Pages/Reporte/Reporte";
+import "./index.css";
+import {
+  createBrowserRouter,
+  RouterProvider,
+  Navigate,
+} from "react-router-dom";
+import ListadoReportes from "./Pages/Reporte/ListadoReportes";
+import InfoReporte from "./Pages/Reporte/InfoReportes";
+import Login from "./Pages/Login";
+import { UserProvider, useUser } from "./Pages/UserContext"; // Importa UserProvider y useUser
+import Router from "./routes/routes";
 
-export const LabsContext = React.createContext<[LabReporte[], React.Dispatch<React.SetStateAction<LabReporte[]>>]>([[], () => {}]);
+export const LabsContext = React.createContext<
+  [LabReporte[], React.Dispatch<React.SetStateAction<LabReporte[]>>]
+>([[], () => {}]);
 
 export interface LabReporte {
   Id: number;
@@ -33,49 +40,18 @@ export interface LabReporte {
   return username ? element : <Navigate to="/Login" replace />;
 };*/
 
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <App /> /* (
-      <ProtectedRoute element={<App />} />
-    ) */
-  },
-  {
-    path: "/Reporte/:Nombre/:Id",
-    element: <Reporte /> /* (
-      <ProtectedRoute element={<Reporte />} />
-    ) */
-  },
-  {
-    path: "/ListadoReporte/:Nombre/:Id",
-    element: <ListadoReportes /> /* (
-      <ProtectedRoute element={<ListadoReportes />} />
-    ) */
-  },
-  {
-    path: "/InfoReporte/:IdReporte/:Nombre/:Id",
-    element: <InfoReporte /> /* (
-      <ProtectedRoute element={<InfoReporte />} />
-    ) */
-  },
-  {
-    path: "/Login",
-    element: <Login />
-  }
-]);
-
 function Root() {
   const [Labs, setLabs] = useState<LabReporte[]>([]);
   return (
     <LabsContext.Provider value={[Labs, setLabs]}>
       <UserProvider>
-        <RouterProvider router={router}></RouterProvider>
+        <Router />
       </UserProvider>
     </LabsContext.Provider>
   );
 }
 
-ReactDOM.createRoot(document.getElementById('root')!).render(
+ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <Root />
   </React.StrictMode>
